requirements.txt
beautifulsoup4==4.12.2
cairocffi @ file:///C:/Users/Gorkee/pipwin/cairocffi-1.3.0-cp311-cp311-win_amd64.whl#sha256=3db8f405cc93d3326d867a7482c3f732bd397cb8cfa696c6a4e926c89f7bd9a9
CairoSVG==2.7.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
chess==1.10.0
cssselect2==0.7.0
defusedxml==0.7.1
docopt==0.6.2
idna==3.6
Js2Py==0.74
packaging==23.2
Pillow==10.1.0
pipwin==0.5.2
pycparser==2.21
pyjsparser==2.7.1
PyPrind==2.11.3
pySmartDL==1.3.4
python-chess==1.999
requests==2.31.0
six==1.16.0
soupsieve==2.5
tinycss2==1.2.1
tzdata==2023.3
tzlocal==5.2
urllib3==2.1.0
webencodings==0.5.1
